Add shadow objects to back stack Values
In an expression like d = a*b + c, the a*b is a temporary that gets
destroyed after that statement executes. For backprop, we need that
object to persist. So we create shadow objects on the heap that back the
objects on the stack. These shadow objects persist till the end of the
program, and will allow correct backprop later.

Lots of subtle C++ issues to make note of! See code comments.