Fatal flaw in design due to ref members!
I wanted to retain ref members in the Value class to permit 1:1
equivalence to Python. However this results in a fatal flaw: in
code like the following:

Value a{2.0, "a1"};
a = Value{3.0, "a2"};

the MAO is invoked on a, which changes the _spv ptr to point to the new
backing object. But the refs are still pointing to the members in the
old destroyed object - this means dangling refs!

Will have to abandon this approach in favor of access fns that
return a ref. Now _that_ ref is newly created on every call, and accessing
the members after changing the _spv ptr will correctly return the updated values.