BUG: UB from direct translation of Python _backward
I was suspicious about capture lifetimes, and Grok set me right. Cannot
capture [this] by value in operator+ since the "this" object could be
temporary in an expression like a*b+c. It will get destroyed at the end
of that statement, and the pointer captured in the lambda is dangling!

Must instead capture the SPs by value and use those within the lambda.

Side note: the capture clause [&,this] is illegal. The intent was to
capture everything by ref except "this" by value. However this is not
allowed! Only legal ways to capture "this" are:
[this], [=], [&], and [*this].