Add autograd support
Fix the UB by capturing SPs to the backing _Value objects. There was
still a subtle bug: returning the lambda accessor by type auto resulted
in a _copy_ of _spv._backward being returned. The set value was going to
a temporary and getting discarded; the original lambda remained
unchanged. Ret type for the accessor must be auto&!