Implement backward pass with cached topo sort
Grok pointed out (correctly) that a static topo array implies that only
one graph can ever be used at a time. Cannot train multiple networks
using the same Value class!

However, I also don't want to have to redo the topo sort on every
backward pass. In a training run the network is not going to change, and
it makes no sense to rerun the topo sort every time. So the graph must be
stored in the root node for that graph. The topo sort is done once the
first time that the backward pass is run; thereafter the cached sort
order is used.