Fix gradients to accumulate
This is to allow correct gradient update in cases like

b = a + a;

Here naive assignment of gradient results in node a getting grad 1.0.
But in this case both sides of the operator are the same object. So we
must += the grad instead of just =.

Similarly for

d = a * b;
e = a + b;
f = d * e;

Here the second pass in backprop from e overwrites the gradients from
the first one (backprop from d). Again fixed by +=.