Change shared_ptr captures in _backward lambdas to raw ptrs
No reason to have shared_ptr captures by value in the _backward lambdas
during graph construction. That incurs a needless refcount increment bc
of the copy, and a decrement at graph destruction. Although this doesn't
affect the backprop hot path, there's really no reason to copy the SPs
because we are not trying to impact ownership of the parent/output
nodes. Changed the capture to take raw ptrs.