Add test to compare Neuron output with PyTorch
For this, first needed to add a facility to inject external weights/bias
into the Neuron. Also added fn call operator with initializer_list
parameter to enable Pythonic out = n({1.0, 2.0, 3.0}) syntax.
TorchNeuron class implements the Neuron functionality using PT infra.

Critical point to note is that this needs to be built with
RelWithDebInfo config. Without that, the register_module call in
TorchNeuron ctor segfaults. Per Gemini this is because the memory layout
of Standard Library containers (like std::string, std::vector, and
std::unordered_map) changes depending on whether you are compiling in
Debug mode or Release mode. The application passes a std::string (the
name "linear") to LibTorch. LibTorch expects the "Release layout" of the
string, but your app provides the "Debug layout." When LibTorch tries to
hash this string to put it in the module dictionary, it reads invalid memory,
resulting in a segfault.

RelWithDebInfo is a hybrid config that preserves the Release ABI (which
plays nice with the Release libtorch I'm using) and also enables
debugging info. With this, the segfault is gone and everything runs
fine.