Change backward pass to use C++20 ranges::for_each
A seemingly simple and straightforward change resulted in a remarkable
amount of headache. Turns out you cannot just use &_Value::_backward in
the call because that is a ptr to a data member. When std::invoke sees
a pointer to a data member, its defined behavior is to access that member,
not call it. Workaround is to use a lambda to call the member lambda.