BUG: cannot compose transform_views in MLP
Added MLP code, but ran into a compile fail with my naive implementation
of operator(). I tried to return a transform_view of Layers. However,
the transform calls Layer::operator() which itself is a transform_view
over Neurons. So the final output of the MLP is a transform_view of
transform_views, which does not yield Values when iterated.

The key difference is that Layers in an MLP are sequential, which means
they need to be evaluated in succession. In contrast, Neurons in a Layer
can be evaluated independently. The correct implementation of
MLP::operator() needs to loop over the Layers computing each output and
feeding into the next Layer.