Test code to compare gradients with PT after backward pass
Another major vibe code-a-thon with Gemini to read out Pytorch MLP
gradients in my neuron-major order, and finally to compare gradient of
each parameter with my MLP. This last loop is a workaround; the version
that used an iterator to the mlp.parameters() join_view strangely
segfaults when the catch2 CHECK_THAT clause is in the loop. Needs
investigation.