Bugfix for intermittent segfaults
The most significant bugfix was to Layer::operator(). Returning a view
was resulting in random segfaults from dangling refs: the transform lambda
captured a ref to the input_range when operator() was called. But that ref
ceased to be valid when iterating over the view later, so it was resulting
in random crashes depending on what preceding/succeeding code was uncommented.
Materializing a vector<Value> fixes this. Gemini FTW!

Added std::formatter for class Value (also courtesy Gemini) so that Values
can be printed out using std::print*.

Final loss computation from the MLP also works!