Bugfix: gradients need to be zeroed before backprop
All backprop ops do a += on the gradient. So the gradients just keep
getting added on each successive backprop. Need to be zeroed out before
every backward pass!

Now the convergence is much slower, and learning rate has to be
increased in order for the loss to converge in a reasonable number of
iterations.