First step towards compiled graph
Create static vec<vec<Value>> to hold input Values, and do a dummy first
fwd pass to build the graph. Subsequent fwd passes just play back the
tape from the topo sort and run the computation with updated param data.
The nodes corresponding to the pre-prepared inputs will need to be
stored in a vec of _Value* ptrs so that data can be updated in-place in
the static graph if needed.